Fails if target table does not exist
The iptables wrapper script does not fail on errors via `set -e` or
similar, because failing in the middle of the run could leave the system
is a bad a state. The iptables commands do write to stderr, however,
providing informative error messages that are mostly ignored by the
role.

Similar to how we're already detecting "unknown option" and failing when
found, we'll also fail if iptables reports "Table does not exist",
which is definitely something an administrator will want to know about,
and certainly constitutes a failure mode.

Closes #24.